+++ packages/orq-rc/RELEASES.md
@@ -9339,3 +9339,13 @@ Based on:
 - [typescript v4.10.0-rc.20] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.20 - packages/orq-rc
+
+## 2026-05-20 07:49:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.3 (2.884.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.21 - packages/orq-rc